requirements.txt
absl-py==2.2.2
aim==3.28.0
aim-ui==3.28.0
aimrecords==0.0.7
aimrocks==0.5.2
aiofiles==24.1.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
base58==2.0.1
bokeh==3.7.2
boto3==1.37.37
botocore==1.37.37
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chex==0.1.86
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
dm-env==1.6
dm-tree==0.1.9
equinox==0.11.4
etils==1.12.2
farama-notifications==0.0.4
fastapi==0.115.12
filelock==3.18.0
flashbax==0.1.2
flax==0.8.5
fonttools==4.57.0
fsspec==2025.3.2
glfw==2.9.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
humanize==4.12.2
idna==3.10
imageio==2.37.0
importlib-resources==6.5.2
jaxopt==0.8.3
jaxtyping==0.3.1
jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.5.1
msgpack==1.1.0
mujoco==3.2.4
mujoco-mjx==3.2.4
narwhals==1.35.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
opt-einsum==3.4.0
optax==0.2.3
orbax-checkpoint==0.5.23
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pyopengl==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
restrictedpython==8.0
rich==14.0.0
s3transfer==0.11.5
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.2
tensorstore==0.1.73
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
trimesh==4.6.8
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wadler-lindig==0.1.5
websockets==15.0.1
wrapt==1.17.2
xyzservices==2025.1.0
zipp==3.21.0